+++ src/BattMo.jl
@@ -230,6 +230,7 @@ include("simulation/simulation_utils.jl")
 include("matlab_interface/matlab_model_setup.jl")
 # include("old_model_setup/model_setup.jl")
 
+include("plotting/makie_ext.jl")
 include("plotting/3d.jl")
 include("plotting/1d.jl")
